Create RegisterContainer.tsx
