+++ src/features/common/Header/index.tsx
@@ -1,11 +1,7 @@
-import { Button } from "@mui/material"
-import { useState } from "react"
-import AlertDialogSlide from "src/components/mui/AlertDialogSlide"
 import { styled } from "styled-components"
-import RegisterForm from "src/features/common/Header/RegisterForm"
-import { RegisterUserPayload } from "src/features/common/users/types/net"
-import UserLogin from "src/features/common/Header/UserLogin"
 import { CookiesProvider } from 'react-cookie';
+import RegisterContainer from "src/features/common/Header/UserLogin/RegisterContainer"
+import LoginContainer from "src/features/common/Header/UserLogin/LoginContainer"
 
 const Wrapper = styled.div`
   display: flex;
@@ -18,40 +14,12 @@ const Wrapper = styled.div`
 `
 
 const Header = () => {
-  const [isDialogOpen, setIsDialogOpen] = useState(false)
-
-  const [registerRequestError, setRegisterRequestError] = useState<RegisterUserPayload>({
-    email: '',
-    username: '',
-    password: ''
-  })
-
-  const onOpenRegisterDialog = () => {
-    setIsDialogOpen(true)
-  }
-
-  const handleCloseDialogData = () => {
-    setIsDialogOpen(false)
-  }
-
-  return <Wrapper>
-    <CookiesProvider>
-      <UserLogin />
-    </CookiesProvider>
-    <Button variant="contained" size="small" onClick={() => onOpenRegisterDialog()}>註冊</Button>
-    <AlertDialogSlide
-      title="註冊"
-      isDialogOpen={isDialogOpen}
-      shouldHideButtons
-      onClose={handleCloseDialogData}
-    >
-      <RegisterForm
-        registerRequestError={registerRequestError}
-        setRegisterRequestError={(newError) => setRegisterRequestError({...registerRequestError, ...newError})}
-        onClose={handleCloseDialogData}
-      />
-    </AlertDialogSlide>
-  </Wrapper>
+  return <CookiesProvider>
+    <Wrapper>
+      <RegisterContainer />
+      <LoginContainer />
+    </Wrapper>
+  </CookiesProvider>
 }
 
 export default Header
